// src/app/models/missionDetails.ts
export interface MissionDetails{
    flight_number: number;
    mission_name: string;
    launch_year: number;
    details: string;
    mission_patch_small: string;
    rocket_name: string;
    rocket_type: string;
    launch_site_name:string;
    article_link: string;
    wikipedia: string;
    video_link: string;
}